Update cis_5_4_4.pp
Fixed umasks being appended to end of files instead of replacing multiple values in if statements when present.  Manual spacing is used as file_line does not allow regex or captures in file statement.  Added login.defs and csh.cshrc as benchmark states '.profile , .bashrc , etc', and 'A default umask setting of 077 causes files and directories created by users to not be readable by any other user on the system'.  Tested.