Disable IPv6 correctly on Debian flavors
There is a rules::ensure_ipv6_is_disabled class that assumes a Red Hat
like system and there is a disable_ipv6 that works for most Linux
systems. The CIS benchmark for Debian9 and Debian10 use the title
'Disable IPv6' and the Ubuntu18 document uses the title 'Ensure IPv6 is
disabled'. All 3 give remediation code that is consistent with our
rules::disable_ipv6 code.