Improved ensure_gdm_login_banner_is_configured
Initially I noticed this rule fail on a Centos7 server. This was caused
by the package libgnome-keyring that was installed. First naive fix was
to change the fact to 'rpm -qa | grep ^gnome', but it should really
check for the package gdm.
This is because the directory /etc/dconf/db/gdm.d/ is provided by this
package. So, if this package is not present, managing
/etc/dconf/db/gdm.d/01-banner-message fails
)secure_linux_cis::rules::ensure_gdm_login_banner_is_configured).

Also removed
manifests/rules/ensure_gdm_login_banner_is_configured_scored.pp, as it
is not used at all?