Ensure /tmp is configured
The previous version of this rule did not work, as the augeas lens for
/etc/fstab expects a different approach for fstab options (no commas or
equal signs are allowed as is)
It also did not directly implement the rule as it is checked by the CIS
CAT tool.

The text of this rule is vague, and speaks of having /tmp on it's own
file system, with one of the reasons so that the noexec option can be
set.

The actual check that the CIS CAT Assessor runs is just a test if /tmp
is a filesystem.

Most filesystem options listed in the documentation for this rule
(noexec, nosuid, nodev) are already implemented in different rules.

Making mode=1777 work requires quite some rework of
secure_linux_cis::mount_options. The only other mount options listed are
the mutual exclusive relatime and strictatime options.