Ensure iptables is not enabled for RedHat8
On CentOS8 and RHEL8, the iptables services is split into a separate
package, iptables-services.

If this package is not installed, then the iptables service is not
present. This gives an error in Puppet, as Puppet tries to mask the
(nonexistent) service iptables.
If the package iptables-services is not present, then there's no need to
mask the service.